Refactored circular dependency
