+++ backend/src/app.py
@@ -8,14 +8,13 @@ from mlutil import get_model, classify_audio_file
 from Spectrogram import sendMelSpectrogram, sendBasicSpectrogram
 from routers.security import get_current_user, LightWeightUser
 from routers.security import router as security_router
+from standaloneops import classify_and_save
 import dao
 import mqtt
 import dao as dao
 import os
 import io
-import asyncio
 import dotenv
-import ssl
 
 from datetime import datetime, timedelta
 
@@ -116,26 +115,6 @@ async def audio_slice_all(current_user: Annotated[LightWeightUser, Depends(get_c
 @app.get(path="/api/audioslices/{asid}")
 async def audio_slice_get(asid: int, current_user: Annotated[LightWeightUser, Depends(get_current_user)], db=Depends(get_db_connection)):
     return await dao.AudioSlice.get(current_user.auid, asid, db)
-
-
-
-async def classify_and_save(audio, audio_file_id, db, model):
-    classifier_output = classify_audio_file(audio, model)
-    slice_insert_tasks = []
-    for classified_slice_name, classified_slice in classifier_output.items():
-        classified_slice["starttime"] = classified_slice.pop("start_time")
-        classified_slice["endtime"] = classified_slice.pop("end_time")
-        slice_insert_tasks.append(
-            asyncio.create_task(
-                dao.AudioSlice.insert(db, audio_file_id, **classified_slice),
-                name=classified_slice_name,
-            )
-        )
-
-    done, pending = await asyncio.wait(slice_insert_tasks)
-    results = map(lambda task: task.result(), done)
-    db.commit()
-    return results
 
 
 @app.post(path="/api/audio/insert")
